fix(test/ec2-lifecyle): add deterministic `TestID` filtering for EC2 instances to prevent test environment collisions on AWS
This commit implements a deterministic TestID for EC2 test instances
based on the current user and git branch. The TestID is then applied as
a tag when provisioning 'eddings' and 'jordan-u' EC2 instances, saved to
a test-session.json file, and is loaded from that file during teardown
and used to identify and remove only those specific test instances.

Changes include:

  - Computing a TestID in `provision.yml` using `whoami` + current git
    branch.
  - Fallback to a whoami + UUID if git info is unavailable.
  - Tagging EC2 instances with the TestID during provisioning.
  - Writing the TestID to `test/test-session.json` for reuse in teardown.yml.
  - Updating `teardown.yml` to filter EC2 instances by TestID.
  - Instructs git to ignore `test/test-session.json`.

Fixes #60